Moved getSecretData method to kube client
